portal-instances-web: uses modal buttons with custom onClick handler
