+++ modules/apps/portal-instances/portal-instances-web/src/main/resources/META-INF/resources/add_instance.jsp
@@ -69,12 +69,6 @@
 
 					</aui:select>
 				</c:if>
-
-				<liferay-frontend:edit-form-footer>
-					<liferay-frontend:edit-form-buttons
-						submitLabel="add"
-					/>
-				</liferay-frontend:edit-form-footer>
 			</div>
 		</div>
 
@@ -85,7 +79,3 @@
 		</div>
 	</liferay-frontend:edit-form>
 </clay:container-fluid>
-
-<liferay-frontend:component
-	module="js/AddInstance"
-/>
