+++ modules/apps/portal-instances/portal-instances-web/src/main/resources/META-INF/resources/css/main.scss
@@ -1,8 +1,4 @@
 .portlet-portal-instances {
-	.dialog-footer {
-		position: fixed;
-	}
-
 	.add-instance-loading {
 		margin-top: 30vh;
